requirements.txt
fastapi
uvicorn
pydantic
python-frontmatter
google-generativeai
openai
anthropic
supabase
python-dotenv
rich
huggingface_hub
tiktok-uploader
moviepy
gTTS
edge-tts
stripe
email-validator
playwright
nest_asyncio
